refactor(#63): redux selector, filename 변경
- redux selector
- filename 변경
